move more json files
