+++ src/common/test-utils/fixtures/branch-delete.json
