+++ src/common/test-utils/fixtures/pull-request-remove-keys.json
