+++ src/common/test-utils/fixtures/pull-request-triggered-by-bot.json
